# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Edu Game AI Service",
    "description": "Async educational game content generation service",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/documents/upload": {
      "post": {
        "tags": [
          "admin-documents"
        ],
        "summary": "Upload System Document",
        "description": "Upload a system-scoped document and trigger Vertex AI Search import.",
        "operationId": "upload_system_document_api_v1_admin_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_system_document_api_v1_admin_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/documents/{doc_id}": {
      "delete": {
        "tags": [
          "admin-documents"
        ],
        "summary": "Delete System Document",
        "description": "Delete a system document from GCS and Firestore tracking.",
        "operationId": "delete_system_document_api_v1_admin_documents__doc_id__delete",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_system_document_api_v1_admin_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          }
        },
        "type": "object",
        "required": [
          "file",
          "subject"
        ],
        "title": "Body_upload_system_document_api_v1_admin_documents_upload_post"
      },
...
      "DocumentResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "file_format": {
            "type": "string",
            "title": "File Format"
          },
          "gcs_uri": {
            "type": "string",
            "title": "Gcs Uri"
          },
          "index_status": {
            "type": "string",
            "title": "Index Status",
            "default": "indexing"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": "user"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "user_id",
          "filename",
          "file_format",
          "gcs_uri"
        ],
        "title": "DocumentResponse",
        "description": "Response after document upload."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}